requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.9.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==10.0.0
protobuf==3.20.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.32.1
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
